Move locator to api.
