+++ crates/warpgate-api/Cargo.toml
@@ -13,6 +13,7 @@ anyhow = { workspace = true }
 schematic = { workspace = true, optional = true, features = ["schema"] }
 serde = { workspace = true }
 serde_json = { workspace = true }
+thiserror = { workspace = true }
 
 [features]
 default = []
