+++ crates/warpgate-api/src/lib.rs
@@ -1,10 +1,12 @@
 mod host;
 mod host_funcs;
+mod locator;
 mod virtual_path;
 
 pub use anyhow::anyhow;
 pub use host::*;
 pub use host_funcs::*;
+pub use locator::*;
 pub use virtual_path::*;
 
 /// Wrap a struct with common derives and serde required attributes.
