+++ crates/warpgate/src/lib.rs
@@ -5,7 +5,6 @@ mod helpers;
 pub mod host_funcs;
 mod id;
 mod loader;
-mod locator;
 mod plugin;
 pub mod test_utils;
 
@@ -14,9 +13,8 @@ pub use error::*;
 pub use helpers::*;
 pub use id::*;
 pub use loader::*;
-pub use locator::*;
 pub use plugin::*;
 
 pub use extism::{Manifest as PluginManifest, Wasm};
 pub use warpgate_api as api;
-pub use warpgate_api::VirtualPath;
+pub use warpgate_api::{GitHubLocator, PluginLocator, PluginLocatorError, VirtualPath};
